+++ docs/source/conf.py
@@ -10,8 +10,8 @@ sys.path.insert(0, os.path.abspath('../src'))
 project = 'AMQCSLdb'
 copyright = '2025, FieryIceStickie'
 author = 'FieryIceStickie'
-version = '1.0'
-release = '1.1.0'
+version = '1.1'
+release = '1.1.1'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration